[IMP] clouder: Split model files by name * Isolate model files by model name
