+++ clouder/models/application.py
@@ -1,151 +1,12 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-# Author: Yannick Buron
-# Copyright 2015, TODAY Clouder SASU
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License with Attribution
-# clause as published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License with
-# Attribution clause along with this program. If not, see
-# <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Copyright 2015 Clouder SASU
+# License LGPL-3.0 or later (http://gnu.org/licenses/lgpl.html).
 
 from datetime import datetime
 import os.path
 import re
 
 from openerp import models, fields, api
-
-from .model import generate_random_password
-
-
-class ClouderApplicationTag(models.Model):
-
-    _name = 'clouder.application.tag'
-
-    name = fields.Char('Name', required=True)
-    application_ids = fields.Many2many(
-        'clouder.application', 'clouder_application_tag_rel',
-        'tag_id', 'application_id', 'Applications')
-
-
-class ClouderApplicationType(models.Model):
-    """
-    Define the application.type object, mainly used to know which python code
-    shall be used for an application.
-
-    For example, when we deploy a base with an Odoo application with 'odoo' as
-    the application type, clouder will check the module clouder_template_odoo
-    to know the specific code to execute.
-    """
-
-    _name = 'clouder.application.type'
-
-    name = fields.Char('Name', required=True)
-    system_user = fields.Char('System User', required=True)
-    localpath = fields.Char('Localpath')
-    localpath_services = fields.Char('Localpath Services')
-    option_ids = fields.One2many(
-        'clouder.application.type.option', 'application_type_id', 'Options'
-    )
-    application_ids = fields.One2many(
-        'clouder.application', 'type_id', 'Applications'
-    )
-    symlink = fields.Boolean('Use Symlink by default?')
-    multiple_databases = fields.Char('Multiples databases?')
-    tag_ids = fields.Many2many(
-        'clouder.application.tag', 'clouder_application_type_tag_rel',
-        'type_id', 'tag_id', 'Tags')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(name)', 'Name must be unique!'),
-    ]
-
-    _order = 'name'
-
-    @api.multi
-    @api.constrains('name', 'system_user')
-    def _check_forbidden_chars_name_sys_user(self):
-        """
-        Check that the application type name does not contain any forbidden
-        characters.
-        """
-        if not all([re.match(r"^[\w\d-]*$", self.name),
-                    re.match(r"^[\w\d-]*$", self.system_user)]):
-            self.raise_error(
-                "Name and system_user can only contain letters, "
-                "digits and -"
-            )
-
-
-class ClouderApplicationTypeOption(models.Model):
-    """
-    Define the application.type.option object, used to know which option can be
-    assigned to application/container/service/base.
-    """
-
-    _name = 'clouder.application.type.option'
-
-    application_type_id = fields.Many2one(
-        'clouder.application.type',
-        'Application Type', ondelete="cascade", required=True)
-    name = fields.Char('Name', required=True)
-    type = fields.Selection(
-        [('application', 'Application'), ('container', 'Container'),
-         ('service', 'Service'), ('base', 'Base')], 'Type', required=True)
-    application_code = fields.Char('Application Code')
-    tag_ids = fields.Many2many(
-        'clouder.application.tag', 'clouder_application_type_option_tag_rel',
-        'option_id', 'tag_id', 'Tags')
-    auto = fields.Boolean('Auto?')
-    required = fields.Boolean('Required?')
-    default = fields.Text('Default value')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(application_type_id,name)',
-         'Options name must be unique per apptype!'),
-    ]
-
-    @api.multi
-    def generate_default(self):
-        res = ''
-        if self.name == 'db_password':
-            res = generate_random_password(20)
-        if self.name == 'secret':
-            res = generate_random_password(50)
-        if self.name == 'ssh_privatekey':
-            res = self.env['clouder.server']._default_private_key()
-        if self.name == 'ssh_publickey':
-            res = self.env['clouder.server']._default_public_key()
-        return res
-
-    @property
-    def get_default(self):
-        if self.default:
-            return self.default
-        else:
-            return self.generate_default()
-
-
-class ClouderApplicationTemplate(models.Model):
-    """
-    """
-
-    _name = 'clouder.application.template'
-
-    name = fields.Char('Name', required=True)
-    link_ids = fields.One2many('clouder.application.link', 'template_id',
-                               'Links')
 
 
 class ClouderApplication(models.Model):
@@ -395,114 +256,3 @@ class ClouderApplication(models.Model):
                         ('template_id', '=', template.id)]):
                     link.reset_template(records=[self])
         return res
-
-
-class ClouderApplicationOption(models.Model):
-    """
-    Define the application.option object, used to define custom values specific
-    to an application.
-    """
-
-    _name = 'clouder.application.option'
-
-    application_id = fields.Many2one('clouder.application', 'Application',
-                                     ondelete="cascade", required=False)
-    template_id = fields.Many2one(
-        'clouder.application.template', 'Template', ondelete="cascade")
-    name = fields.Many2one('clouder.application.type.option', 'Option',
-                           required=True)
-    value = fields.Text('Value')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(application_id,template_id,name)',
-         'Option name must be unique per application!'),
-    ]
-
-
-class ClouderApplicationLink(models.Model):
-    """
-    Define the application.link object, used to know which others applications
-    can be link to this application.
-    """
-
-    _name = 'clouder.application.link'
-
-    _inherit = ['clouder.template.one2many']
-
-    _template_parent_model = 'clouder.application'
-    _template_parent_many2one = 'application_id'
-    _template_fields = ['required', 'auto', 'make_link', 'container', 'base']
-
-    application_id = fields.Many2one('clouder.application', 'Application',
-                                     ondelete="cascade", required=False)
-    template_id = fields.Many2one(
-        'clouder.application.template', 'Template',
-        ondelete="cascade", required=False)
-    name = fields.Many2one('clouder.application', 'Application', required=True)
-    required = fields.Boolean('Required?')
-    auto = fields.Boolean('Auto?')
-    make_link = fields.Boolean('Make docker link?')
-    container = fields.Boolean('Container?')
-    base = fields.Boolean('Base?')
-    next = fields.Many2one('clouder.container', 'Next')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(application_id,template_id,name)',
-         'Links must be unique per application!'),
-    ]
-
-
-class ClouderApplicationMetadata(models.Model):
-    """
-    Defines an object to store metadata linked to an application
-    """
-
-    _name = 'clouder.application.metadata'
-
-    application_id = fields.Many2one(
-        'clouder.application', 'Application',
-        ondelete="cascade", required=True)
-    name = fields.Char('Name', required=True, size=64)
-    clouder_type = fields.Selection(
-        [
-            ('container', 'Container'),
-            ('base', 'Base')
-        ], 'Type', required=True)
-    is_function = fields.Boolean(
-        'Function', help="Is the value computed by a function?",
-        required=False, default=False)
-    func_name = fields.Char('Function Name', size=64)
-    default_value = fields.Text('Default Value')
-    value_type = fields.Selection(
-        [
-            ('int', 'Integer'),
-            ('float', 'Float'),
-            ('char', 'Char')
-        ], 'Data Type', required=True)
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(application_id,name, clouder_type)',
-         'Metadata must be unique per application!'),
-    ]
-
-    @api.depends('is_function', 'func_name')
-    @api.multi
-    def _check_function(self):
-        """
-        Checks that the function name is defined
-        and exists if is_functions is set to True
-        """
-        for metadata in self:
-            if metadata.is_function:
-                if not metadata.func_name:
-                    self.raise_error(
-                        "You must enter the function name "
-                        "to set is_function to true."
-                    )
-                else:
-                    obj_env = self.env['clouder.'+self.clouder_type]
-                    if not getattr(obj_env, self.func_name, False):
-                        self.raise_error(
-                            "Invalid function name %s for clouder.base",
-                            self.name.func_name,
-                        )
