+++ clouder/models/base.py
@@ -1,24 +1,6 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-# Author: Yannick Buron
-# Copyright 2015, TODAY Clouder SASU
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License with Attribution
-# clause as published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License with
-# Attribution clause along with this program. If not, see
-# <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Copyright 2015 Clouder SASU
+# License LGPL-3.0 or later (http://gnu.org/licenses/lgpl.html).
 
 
 from datetime import datetime, timedelta
@@ -27,57 +9,10 @@ import re
 
 from openerp import models, fields, api
 
-from .model import generate_random_password
+from ..tools import generate_random_password
 
 
 _logger = logging.getLogger(__name__)
-
-
-class ClouderDomain(models.Model):
-    """
-    Define the domain object, which represent all domains which can be linked
-    to the bases hosted in this clouder.
-    """
-
-    _name = 'clouder.domain'
-    _inherit = ['clouder.model']
-
-    name = fields.Char('Domain name', required=True)
-    organisation = fields.Char('Organisation', required=True)
-    dns_id = fields.Many2one('clouder.container', 'DNS Server', required=False)
-    cert_key = fields.Text('Wildcard Cert Key')
-    cert_cert = fields.Text('Wildcart Cert')
-    public = fields.Boolean('Public?')
-    partner_id = fields.Many2one(
-        'res.partner', 'Manager',
-        default=lambda self: self.env.user.partner_id)
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(name)', 'Name must be unique!'),
-    ]
-
-    @api.multi
-    @api.constrains('name')
-    def _check_name(self):
-        """
-        Check that the domain name does not contain any forbidden
-        characters.
-        """
-        if not re.match(r"^[\w\d.-]*$", self.name):
-            self.raise_error(
-                "Name can only contains letters, digits - and dot"
-            )
-
-    @api.multi
-    def write(self, vals):
-
-        if 'dns_id' in vals:
-            self.purge()
-
-        super(ClouderDomain, self).write(vals)
-
-        if 'dns_id' in vals:
-            self.deploy()
 
 
 class ClouderBase(models.Model):
@@ -962,287 +897,3 @@ class ClouderBase(models.Model):
         Renew a certificate
         """
         return True
-
-
-class ClouderBaseOption(models.Model):
-    """
-    Define the base.option object, used to define custom values specific
-    to a base.
-    """
-    _name = 'clouder.base.option'
-
-    base_id = fields.Many2one('clouder.base', 'Base', ondelete="cascade",
-                              required=True)
-    name = fields.Many2one('clouder.application.type.option', 'Option',
-                           required=True)
-    value = fields.Text('Value')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(base_id,name)',
-         'Option name must be unique per base!'),
-    ]
-
-    @api.multi
-    @api.constrains('base_id')
-    def _check_required(self):
-        """
-        Check that we specify a value for the option
-        if this option is required.
-        """
-        if self.name.required and not self.value:
-            self.raise_error(
-                'You need to specify a value for the option "%s" '
-                'for the base "%s".',
-                self.name.name, self.base_id.name,
-            )
-
-
-class ClouderBaseLink(models.Model):
-    """
-    Define the base.link object, used to specify the applications linked
-    to a base.
-    """
-    _name = 'clouder.base.link'
-    _inherit = ['clouder.model']
-    _autodeploy = False
-
-    base_id = fields.Many2one('clouder.base', 'Base', ondelete="cascade",
-                              required=True)
-    name = fields.Many2one('clouder.application', 'Application',
-                           required=True)
-    target = fields.Many2one('clouder.container', 'Target')
-    required = fields.Boolean('Required?')
-    auto = fields.Boolean('Auto?')
-    deployed = fields.Boolean('Deployed?', readonly=True)
-
-    @property
-    def target_base(self):
-        """
-        Property returning the first base of the target container.
-        """
-        return self.target.base_ids and \
-            self.target.base_ids[0]
-
-    @api.multi
-    @api.constrains('base_id')
-    def _check_required(self):
-        """
-        Check that we specify a value for the link
-        if this link is required.
-        """
-        if self.required and not self.target:
-            self.raise_error(
-                'You need to specify a link to "%s" for the base "%s"',
-                self.name.name, self.base_id.name,
-            )
-
-    @api.multi
-    def deploy_link(self):
-        """
-        Hook which can be called by submodules to execute commands when we
-        deploy a link.
-        """
-        self.purge_link()
-        self.deployed = True
-        return
-
-    @api.multi
-    def purge_link(self):
-        """
-        Hook which can be called by submodules to execute commands when we
-        purge a link.
-        """
-        self.deployed = False
-        return
-
-    def control(self):
-        """
-        Make the control to know if we can launch the deploy/purge.
-        """
-        if not self.target:
-            self.log(
-                'The target isnt configured in the link, skipping deploy link')
-            return False
-        return True
-
-    @api.multi
-    def deploy_(self):
-        self = self.with_context(no_enqueue=True)
-        self.do(
-            'deploy_link ' + self.name.name, 'deploy_exec', where=self.base_id)
-
-    @api.multi
-    def deploy_exec(self):
-        """
-        Control and call the hook to deploy the link.
-        """
-        self.control() and self.deploy_link()
-
-    @api.multi
-    def purge_(self):
-        self = self.with_context(no_enqueue=True)
-        self.do(
-            'purge_link ' + self.name.name, 'purge_exec', where=self.base_id)
-
-    @api.multi
-    def purge_exec(self):
-        """
-        Control and call the hook to purge the link.
-        """
-        self.control() and self.purge_link()
-
-
-class ClouderBaseChild(models.Model):
-    """
-    Define the base.child object, used to specify the applications linked
-    to a container.
-    """
-
-    _name = 'clouder.base.child'
-    _inherit = ['clouder.model']
-    _autodeploy = False
-
-    base_id = fields.Many2one(
-        'clouder.base', 'Base', ondelete="cascade", required=True)
-    name = fields.Many2one(
-        'clouder.application', 'Application', required=True)
-    sequence = fields.Integer('Sequence')
-    domainname = fields.Char('Name')
-    domain_id = fields.Many2one('clouder.domain', 'Domain')
-    container_id = fields.Many2one(
-        'clouder.container', 'Container')
-    child_id = fields.Many2one(
-        'clouder.container', 'Container')
-    save_id = fields.Many2one('clouder.save',
-                              'Restore this save on deployment')
-
-    _order = 'sequence'
-
-    @api.multi
-    @api.constrains('child_id')
-    def _check_child_id(self):
-        if self.child_id and not self.child_id.parent_id == self:
-            self.raise_error(
-                "The child container is not correctly linked to the parent",
-            )
-
-    @api.multi
-    def create_child(self):
-        self = self.with_context(no_enqueue=True)
-        self.do(
-            'create_child ' + self.name.name,
-            'create_child_exec', where=self.base_id)
-
-    @api.multi
-    def create_child_exec(self):
-        self = self.with_context(autocreate=True)
-        self.delete_child_exec()
-        self.child_id = self.env['clouder.base'].create({
-            'name':
-                self.domainname or self.base_id.name + '-' + self.name.code,
-            'domain_id':
-                self.domainname and
-                self.domain_id or self.base_id.domain_id.id,
-            'parent_id': self.id,
-            'environment_id': self.base_id.environment_id.id,
-            'application_id': self.name.id,
-            'container_id': self.container_id.id
-        })
-        if self.save_id:
-            self.save_id.container_id = self.child_id.container_id
-            self.save_id.base_id = self.child_id
-            self.save_id.restore()
-
-    @api.multi
-    def delete_child(self):
-        self.do(
-            'delete_child ' + self.name.name,
-            'delete_child_exec', where=self.base_id)
-
-    @api.multi
-    def delete_child_exec(self):
-        self.child_id and self.child_id.unlink()
-
-
-class ClouderBaseMetadata(models.Model):
-    """
-    Defines an object to store metadata linked to an application
-    """
-
-    _name = 'clouder.base.metadata'
-
-    name = fields.Many2one(
-        'clouder.application.metadata', 'Application Metadata',
-        ondelete="cascade", required=True)
-    base_id = fields.Many2one(
-        'clouder.base', 'Base', ondelete="cascade", required=True)
-    value_data = fields.Text('Value')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(name, base_id)',
-         'Metadata must be unique per container!'),
-    ]
-
-    @property
-    def value(self):
-        """
-        Property that returns the value formatted by type
-        """
-        def _missing_function():
-            # If the function is missing, raise an exception
-            self.raise_error(
-                'Invalid function name "%s" for "clouder.base".',
-                self.name.func_name,
-            )
-
-        # Computing the function if needed
-        val_to_convert = self.value_data
-        if self.name.is_function:
-            val_to_convert = "{0}".format(getattr(
-                self.base_id, self.name.func_name, _missing_function)())
-            # If it is a function,
-            # the text version should be updated for display
-            self.with_context(skip_check=True).write({
-                'value_data': val_to_convert})
-
-        # Empty value
-        if not val_to_convert:
-            return False
-
-        # value_type cases
-        if self.name.value_type == 'int':
-            return int(val_to_convert)
-        elif self.name.value_type == 'float':
-            return float(val_to_convert)
-        # Defaults to char
-        return str(val_to_convert)
-
-    @api.multi
-    @api.constrains('name')
-    def _check_clouder_type(self):
-        """
-        Checks that the metadata is intended for containers
-        """
-        if self.name.clouder_type != 'base':
-            self.raise_error(
-                "This metadata is intended for %s only.",
-                self.name.clouder_type,
-            )
-
-    @api.multi
-    @api.constrains('name', 'value_data')
-    def _check_object(self):
-        """
-        Checks if the data can be loaded properly
-        """
-        if 'skip_check' in self.env.context and self.env.context['skip_check']:
-            return
-        # call the value property to see if the metadata can be loaded properly
-        try:
-            self.value
-        except ValueError:
-            # User display
-            self.raise_error(
-                'Invalid value for type "%s": \n\t"%s"\n',
-                self.name.value_type, self.value_data,
-            )
