+++ clouder/models/config_settings.py
@@ -1,41 +1,11 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-# Author: Yannick Buron
-# Copyright 2015, TODAY Clouder SASU
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License with Attribution
-# clause as published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License with
-# Attribution clause along with this program. If not, see
-# <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Copyright 2015 Clouder SASU
+# License LGPL-3.0 or later (http://gnu.org/licenses/lgpl.html).
 
 import os.path
 import re
 
 from openerp import models, fields, api
-
-
-class ClouderConfigBackupMethod(models.Model):
-    """
-    Define the config.backup.method object, which represent all backup method
-    available for save.
-    """
-
-    _name = 'clouder.config.backup.method'
-    _description = 'Backup Method'
-
-    name = fields.Char('Name', required=True)
 
 
 class ClouderConfigSettings(models.Model):
