+++ clouder/models/container.py
@@ -1,30 +1,9 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-# Author: Yannick Buron
-# Copyright 2015, TODAY Clouder SASU
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License with Attribution
-# clause as published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License with
-# Attribution clause along with this program. If not, see
-# <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# Copyright 2015 Clouder SASU
+# License LGPL-3.0 or later (http://gnu.org/licenses/lgpl.html).
 
 from openerp import models, fields, api
-from openerp import modules
 
-import os.path
-import socket
 import re
 
 import time
@@ -32,456 +11,6 @@ from datetime import datetime, timedelta
 
 import logging
 _logger = logging.getLogger(__name__)
-
-
-class ClouderOneclick(models.Model):
-
-    _name = 'clouder.oneclick'
-
-    name = fields.Char('Nom', required=True)
-    code = fields.Char('Code', required=True)
-
-
-class ClouderServer(models.Model):
-    """
-    Define the server object, which represent the servers
-    clouder can connect to.
-    """
-
-    _name = 'clouder.server'
-    _inherit = ['clouder.model']
-    _sql_constraints = [
-        ('domain_id_name_uniq', 'unique(domain_id, name)',
-         'This name already exists on this domain.'),
-        ('ip_ssh_port_uniq', 'unique(ip, ssh_port)',
-         'Another server is already setup for this SSH Address and Port.'),
-    ]
-
-    @api.model
-    def _default_private_key(self):
-        """
-        Generate a couple of keys visible use on the server form, which
-        we can easily add to the server to connect it.
-        """
-
-        destroy = True
-        if not self.local_dir_exist(self.get_directory_key()):
-            self._create_key()
-            destroy = False
-
-        key = self.execute_local([
-            'cat', self.get_directory_key('key'),
-        ])
-
-        if destroy:
-            self._destroy_key()
-        return key
-
-    @api.model
-    def _default_public_key(self):
-        """
-        Generate a couple of keys visible use on the server form, which
-        we can easily add to the server to connect it.
-        """
-
-        destroy = True
-        if not self.local_dir_exist(self.get_directory_key()):
-            self._create_key()
-            destroy = False
-
-        key = self.execute_local([
-            'cat', self.get_directory_key('key.pub'),
-        ])
-
-        if destroy:
-            self._destroy_key()
-        return key
-
-    key_file = fields.Char(
-        compute='_compute_key_file',
-        store=True,
-    )
-    name = fields.Char('Prefix', required=True)
-    domain_id = fields.Many2one('clouder.domain', 'Domain', required=True)
-    ip = fields.Char('IP')
-    environment_id = fields.Many2one('clouder.environment', 'Environment',
-                                     required=True)
-    login = fields.Char('Login')
-    ssh_port = fields.Integer('SSH port', default='22')
-
-    private_key = fields.Text(
-        'SSH Private Key',
-        default=lambda s: s._default_private_key(),
-        compute='_compute_private_key',
-        inverse='_inverse_private_key',
-    )
-    public_key = fields.Text(
-        'SSH Public Key',
-        default=lambda s: s._default_public_key(),
-        compute='_compute_public_key',
-        inverse='_inverse_public_key',
-    )
-    start_port = fields.Integer('Start Port', required=True)
-    end_port = fields.Integer('End Port', required=True)
-    public_ip = fields.Boolean(
-        'Assign ports with public ip?',
-        help="This is especially useful if you want to have several "
-             "infrastructures on the same server, by using same ports but "
-             "different ips. Otherwise the ports will be bind to "
-             "all interfaces.")
-    public = fields.Boolean('Public?')
-    supervision_id = fields.Many2one('clouder.container', 'Supervision Server')
-    runner_id = fields.Many2one('clouder.container', 'Runner')
-    salt_minion_id = fields.Many2one(
-        'clouder.container', 'Salt Minion', readonly=True)
-    control_dns = fields.Boolean('Control DNS?')
-    oneclick_ids = fields.Many2many(
-        'clouder.oneclick', 'clouder_server_oneclick_rel',
-        'container_id', 'oneclick_id', 'Oneclick Deployment')
-    oneclick_ports = fields.Boolean('Assign critical ports?')
-
-    @api.multi
-    @api.depends('name', 'domain_id.name')
-    def _compute_key_file(self):
-        for server in self:
-            server.key_file = os.path.join(
-                server.home_directory, '.ssh', 'keys',
-                '%s.%s' % (server.name, server.domain_id.name),
-            )
-
-    @api.multi
-    def _create_key(self):
-        """
-        Generate a key on the filesystem.
-        """
-        if not self.env.ref('clouder.clouder_settings').email_sysadmin:
-            self.raise_error(
-                "You need to specify the sysadmin email in configuration",
-            )
-
-        self.execute_local(['mkdir', self.get_directory_key()])
-        self.execute_local([
-            'ssh-keygen', '-t', 'rsa', '-C', self.email_sysadmin, '-f',
-            self.get_directory_key('key'), '-N', '',
-        ])
-        return True
-
-    @api.multi
-    def _destroy_key(self):
-        """
-        Destroy the key after once we don't need it anymore.
-        """
-        self.execute_local(['rm', '-rf', self.get_directory_key()])
-        return True
-
-    @api.multi
-    @api.depends('private_key', 'key_file')
-    def _compute_private_key(self):
-        for server in self:
-            server.private_key = self.execute_local(['cat', server.key_file])
-
-    @api.multi
-    @api.depends('public_key', 'key_file')
-    def _compute_public_key(self):
-        for server in self:
-            server.public_key = self.execute_local([
-                'cat', '%s.pub' % server.key_file,
-            ])
-
-    @api.multi
-    def _inverse_private_key(self):
-        """
-        """
-        for server in self:
-            name = server.fulldomain
-            self.execute_local([
-                'mkdir', '-p',
-                os.path.join(server.home_directory, '.ssh', 'keys'),
-            ])
-            key_file = os.path.join(
-                server.home_directory, '.ssh', 'keys', name,
-            )
-            self.execute_write_file(
-                key_file, server.private_key, operator='w',
-            )
-            self.execute_local(['chmod', '600', key_file])
-
-    @api.multi
-    def _inverse_public_key(self):
-        """
-        """
-        for server in self:
-            key_dir = os.path.join(server.home_directory, '.ssh', 'keys')
-            self.execute_local(['mkdir', '-p', key_dir])
-            key_file = os.path.join(key_dir, server.fulldomain)
-            key_file_pub = '%s.pub' % key_file
-            self.execute_write_file(
-                key_file_pub, server.public_key, operator='w',
-            )
-            self.execute_local(['chmod', '600', key_file_pub])
-
-    @property
-    def fulldomain(self):
-        """
-        """
-
-        fulldomain = '%s.%s' % (self.name, self.domain_id.name)
-        if self.control_dns and self.domain_id.dns_id:
-            ip = socket.gethostbyname(fulldomain)
-            if ip != self.ip:
-                self.raise_error(
-                    'Could not resolve hostname of the server "%s"',
-                    fulldomain,
-                )
-        return fulldomain
-
-    @api.multi
-    @api.constrains('name', 'ip')
-    def _check_name_ip(self):
-        """
-        Check that the server domain does not contain any forbidden
-        characters.
-        """
-        if not re.match(r"^[\w\d-]*$", self.name):
-            self.raise_error(
-                "Name can only contains letters, digits, -",
-            )
-        if not re.match(r"^[\d:.]*$", self.ip):
-            self.raise_error(
-                "IP can only contains digits, dots and :",
-            )
-
-    @api.multi
-    def deploy_ssh_config(self):
-        for server in self:
-            name = server.fulldomain
-            ssh_config = os.path.join(self.home_directory, '.ssh', 'config')
-            sed = os.path.join(
-                modules.get_module_path('clouder'), 'res', 'sed.sh',
-            )
-            self.execute_local([sed, name, ssh_config])
-            self.execute_write_file(ssh_config, 'Host %s' % name)
-            self.execute_write_file(
-                ssh_config, '\n  HostName %s' % server.ip,
-            )
-            self.execute_write_file(
-                ssh_config, '\n  Port %s' % server.ssh_port,
-            )
-            self.execute_write_file(
-                ssh_config, '\n  User %s' % (server.login or 'root'),
-            )
-            self.execute_write_file(
-                ssh_config, '\n  IdentityFile ~/.ssh/keys/%s' % name,
-            )
-            self.execute_write_file(
-                ssh_config, '\n#END %s\n' % name,
-            )
-
-    @api.multi
-    def write(self, vals):
-        res = super(ClouderServer, self).write(vals)
-        self.deploy_ssh_config()
-        return res
-
-    @api.multi
-    def do(self, name, action, where=False):
-        if action == 'deploy_frame':
-            self = self.with_context(no_enqueue=True)
-        return super(ClouderServer, self).do(name, action, where=where)
-
-    @api.multi
-    def start_containers(self):
-        self = self.with_context(no_enqueue=True)
-        self.do('start_containers', 'start_containers_exec')
-
-    @api.multi
-    def start_containers_exec(self):
-        """
-        Restart all containers of the server.
-        """
-        containers = self.env['clouder.container'].search(
-            [('server_id', '=', self.id)])
-        for container in containers:
-            container.start()
-
-    @api.multi
-    def stop_containers(self):
-        self = self.with_context(no_enqueue=True)
-        self.do('stop_containers', 'stop_containers_exec')
-
-    @api.multi
-    def stop_containers_exec(self):
-        """
-        Stop all container of the server.
-        """
-        containers = self.env['clouder.container'].search(
-            [('server_id', '=', self.id)])
-        for container in containers:
-            container.stop()
-
-    @api.multi
-    def test_connection(self):
-        """
-        Test connection to the server.
-        """
-        self.connect()
-        self.raise_error('Connection successful!')
-
-    @api.multi
-    def deploy(self):
-        """
-        """
-        self.deploy_ssh_config()
-        super(ClouderServer, self).deploy()
-
-    @api.multi
-    def purge(self):
-        """
-        """
-        super(ClouderServer, self).purge()
-
-    @api.multi
-    def deploy_dns(self):
-        self = self.with_context(no_enqueue=True)
-        self.do('deploy_dns %s' % self.fulldomain, 'deploy_dns_exec')
-
-    @api.multi
-    def deploy_dns_exec(self):
-        self.purge_dns_exec()
-
-        if self.domain_id.dns_id:
-            self.domain_id.dns_id.execute([
-                'echo "%s IN A %s" >> "%s"' % (
-                    self.name, self.ip, self.domain_id.configfile,
-                )
-            ])
-            self.domain_id.refresh_serial(self.fulldomain)
-            # self.control_dns = True
-
-    @api.multi
-    def purge_dns(self):
-        self = self.with_context(no_enqueue=True)
-        self.do('purge_dns', 'purge_dns_exec')
-
-    @api.multi
-    def purge_dns_exec(self):
-        self.control_dns = False
-        if self.domain_id.dns_id:
-            self.domain_id.dns_id.execute([
-                'sed', '-i', r'"/%s\sIN\sA/d"' % self.name,
-                self.domain_id.configfile,
-            ])
-            self.domain_id.refresh_serial()
-
-    @api.multi
-    def oneclick_deploy_element(
-            self, type, code, container=False, code_container='', ports=None):
-
-        if not ports:
-            ports = []
-
-        application_obj = self.env['clouder.application']
-        container_obj = self.env['clouder.container']
-        port_obj = self.env['clouder.container.port']
-        base_obj = self.env['clouder.base']
-
-        application = application_obj.search([('code', '=', code)])
-
-        if not container and code_container:
-            container = container_obj.search([
-                ('environment_id', '=', self.environment_id.id),
-                ('suffix', '=', code_container)])
-        if not container:
-            container = container_obj.search([
-                ('environment_id', '=', self.environment_id.id),
-                ('suffix', '=', code)])
-
-        if type == 'container':
-            if not container:
-                # ports = []
-                # if self.oneclick_ports:
-                #     ports = [(0,0,{'name':'bind', 'localport': 53,
-                # 'hostport': 53, 'expose': 'internet', 'udp': True})]
-                container = container_obj.create({
-                    'suffix': code,
-                    'environment_id': self.environment_id.id,
-                    'server_id': self.id,
-                    'application_id': application.id,
-                })
-                if self.oneclick_ports and ports:
-                    for port in ports:
-                        port_record = port_obj.search([
-                            ('container_id', '=', container.childs['exec'].id),
-                            ('localport', '=', port)])
-                        port_record.write({'hostport': port})
-                    container = container.with_context(container_childs=False)
-                    container.childs['exec'].deploy()
-            return container
-
-        if type == 'base':
-            base = base_obj.search([
-                ('name', '=', code), ('domain_id', '=', self.domain_id.id)])
-            if not base:
-                base = base_obj.create({
-                    'name': code,
-                    'domain_id': self.domain_id.id,
-                    'environment_id': self.environment_id.id,
-                    'title': application.name,
-                    'application_id': application.id,
-                    'container_id': container.id,
-                    'admin_name': 'admin',
-                    'admin_password': 'adminadmin',
-                    'ssl_only': True,
-                    'autosave': True,
-                })
-            return base
-
-        if type == 'subservice':
-            if not container_obj.search([
-                    ('environment_id', '=', self.environment_id.id),
-                    ('suffix', '=', container.name + '-test')]):
-                container.reset_base_ids = [
-                    (6, 0, [b.id for b in container.base_ids])]
-                container.subservice_name = 'test'
-                container.install_subservice()
-
-    @api.multi
-    def oneclick_deploy(self):
-        self = self.with_context(no_enqueue=True)
-        self.do('oneclick_deploy', 'oneclick_deploy_exec')
-
-    @api.multi
-    def oneclick_deploy_exec(self):
-        # TODO check that ns record of the domain is the IP
-        return
-
-    @api.multi
-    def oneclick_purge(self):
-        self = self.with_context(no_enqueue=True)
-        self.do('oneclick_purge', 'oneclick_purge_exec')
-
-    @api.multi
-    def oneclick_purge_exec(self):
-        return
-
-    @api.multi
-    def clean(self):
-        self = self.with_context(no_enqueue=True)
-        self.do('clean', 'clean_exec')
-
-    @api.multi
-    def clean_exec(self):
-        """
-        Clean the server from unused containers / images / volumes.
-        http://blog.yohanliyanage.com/2015/05/docker-clean-up-after-yourself/
-        """
-        self.execute(['docker', 'rmi $(docker images -f "dangling=true" -q)'])
-        self.execute(['docker', 'rmi', '-f', '$(docker images -q)'])
-        self.execute([
-            'docker',
-            'run -v /var/run/docker.sock:/var/run/docker.sock '
-            '-v /var/lib/docker:/var/lib/docker '
-            '--rm martin/docker-cleanup-volumes'])
 
 
 class ClouderContainer(models.Model):
@@ -1776,333 +1305,3 @@ class ClouderContainer(models.Model):
                 reset_base_name=subbase_name, reset_container=subservice)
             base.reset_base_exec()
         self.subservice_name = False
-
-
-class ClouderContainerPort(models.Model):
-    """
-    Define the container.port object, used to define the ports which
-    will be mapped in the container.
-    """
-
-    _name = 'clouder.container.port'
-
-    container_id = fields.Many2one(
-        'clouder.container', 'Container', ondelete="cascade", required=True)
-    name = fields.Char('Name', required=True)
-    localport = fields.Char('Local port', required=True)
-    hostport = fields.Char('Host port')
-    expose = fields.Selection(
-        [('internet', 'Internet'), ('local', 'Local')], 'Expose?',
-        required=True, default='local')
-    udp = fields.Boolean('UDP?')
-    use_hostport = fields.Boolean('Use hostpost?')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(container_id,name)',
-         'Port name must be unique per container!'),
-    ]
-
-
-class ClouderContainerVolume(models.Model):
-    """
-    Define the container.volume object, used to define the volume which
-    will be saved in the container or will be linked to a directory
-    in the host server.
-    """
-
-    _name = 'clouder.container.volume'
-
-    container_id = fields.Many2one(
-        'clouder.container', 'Container', ondelete="cascade", required=True)
-    name = fields.Char('Path', required=True)
-    hostpath = fields.Char('Host path')
-    user = fields.Char('System User')
-    readonly = fields.Boolean('Readonly?')
-    nosave = fields.Boolean('No save?')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(container_id,name)',
-         'Volume name must be unique per container!'),
-    ]
-
-
-class ClouderContainerOption(models.Model):
-    """
-    Define the container.option object, used to define custom values
-    specific to a container.
-    """
-
-    _name = 'clouder.container.option'
-
-    container_id = fields.Many2one(
-        'clouder.container', 'Container', ondelete="cascade", required=True)
-    name = fields.Many2one(
-        'clouder.application.type.option', 'Option', required=True)
-    value = fields.Text('Value')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(container_id,name)',
-         'Option name must be unique per container!'),
-    ]
-
-    @api.multi
-    @api.constrains('container_id')
-    def _check_required(self):
-        """
-        Check that we specify a value for the option
-        if this option is required.
-        """
-        if self.name.required and not self.value:
-            self.raise_error(
-                'You need to specify a value for the option '
-                '"%s" for the container "%s".',
-                self.name.name, self.container_id.name,
-            )
-
-
-class ClouderContainerLink(models.Model):
-    """
-    Define the container.link object, used to specify the applications linked
-    to a container.
-    """
-
-    _name = 'clouder.container.link'
-    _inherit = ['clouder.model']
-    _autodeploy = False
-
-    container_id = fields.Many2one(
-        'clouder.container', 'Container', ondelete="cascade", required=True)
-    name = fields.Many2one(
-        'clouder.application', 'Application', required=True)
-    target = fields.Many2one('clouder.container', 'Target')
-    required = fields.Boolean('Required?')
-    auto = fields.Boolean('Auto?')
-    make_link = fields.Boolean('Make docker link?')
-    deployed = fields.Boolean('Deployed?', readonly=True)
-
-    @api.multi
-    @api.constrains('container_id')
-    def _check_required(self):
-        """
-        Check that we specify a value for the link
-        if this link is required.
-        """
-        if self.required and not self.target \
-                and not self.container_id.child_ids:
-            self.raise_error(
-                'You need to specify a link to '
-                '"%s" for the container "%s".',
-                self.name.name, self.container_id.name,
-            )
-
-    @api.multi
-    def deploy_link(self):
-        """
-        Hook which can be called by submodules to execute commands when we
-        deploy a link.
-        """
-        self.purge_link()
-        self.deployed = True
-        return
-
-    @api.multi
-    def purge_link(self):
-        """
-        Hook which can be called by submodules to execute commands when we
-        purge a link.
-        """
-        self.deployed = False
-        return
-
-    @api.multi
-    def control(self):
-        """
-        Make the control to know if we can launch the deploy/purge.
-        """
-        if self.container_id.child_ids:
-            self.log('The container has children, skipping deploy link')
-            return False
-        if not self.target:
-            self.log('The target isnt configured in the link, '
-                     'skipping deploy link')
-            return False
-        return True
-
-    @api.multi
-    def deploy_(self):
-        self = self.with_context(no_enqueue=True)
-        self.do(
-            'deploy_link ' + self.name.name,
-            'deploy_exec', where=self.container_id)
-
-    @api.multi
-    def deploy_exec(self):
-        """
-        Control and call the hook to deploy the link.
-        """
-        self.control() and self.deploy_link()
-
-    @api.multi
-    def purge_(self):
-        self = self.with_context(no_enqueue=True)
-        self.do(
-            'purge_link ' + self.name.name,
-            'purge_exec', where=self.container_id)
-
-    @api.multi
-    def purge_exec(self):
-        """
-        Control and call the hook to purge the link.
-        """
-        self.control() and self.purge_link()
-
-
-class ClouderContainerChild(models.Model):
-    """
-    Define the container.link object, used to specify the applications linked
-    to a container.
-    """
-
-    _name = 'clouder.container.child'
-    _inherit = ['clouder.model']
-    _autodeploy = False
-
-    container_id = fields.Many2one(
-        'clouder.container', 'Container', ondelete="cascade", required=True)
-    name = fields.Many2one(
-        'clouder.application', 'Application', required=True)
-    sequence = fields.Integer('Sequence')
-    server_id = fields.Many2one(
-        'clouder.server', 'Server')
-    child_id = fields.Many2one(
-        'clouder.container', 'Container')
-    save_id = fields.Many2one(
-        'clouder.save', 'Restore this save on deployment')
-
-    _order = 'sequence'
-
-    @api.multi
-    @api.constrains('child_id')
-    def _check_child_id(self):
-        if self.child_id and not self.child_id.parent_id == self:
-            self.raise_error(
-                "The child container is not correctly linked to the parent",
-            )
-
-    @api.multi
-    def create_child(self):
-        self = self.with_context(no_enqueue=True)
-        self.do(
-            'create_child ' + self.name.name,
-            'create_child_exec', where=self.container_id)
-
-    @api.multi
-    def create_child_exec(self):
-        container = self.container_id
-        self = self.with_context(autocreate=True)
-        self.delete_child_exec()
-        self.env['clouder.container'].create({
-            'environment_id': container.environment_id.id,
-            'suffix': container.suffix + '-' + self.name.code,
-            'parent_id': self.id,
-            'application_id': self.name.id,
-            'server_id': self.server_id.id or container.server_id.id
-        })
-        if self.save_id:
-            self.save_id.container_id = self.child_id
-            self.save_id.restore()
-
-    @api.multi
-    def delete_child(self):
-        self = self.with_context(no_enqueue=True)
-        self.do(
-            'delete_child ' + self.name.name,
-            'delete_child_exec', where=self.container_id)
-
-    @api.multi
-    def delete_child_exec(self):
-        self.child_id and self.child_id.unlink()
-
-
-class ClouderContainerMetadata(models.Model):
-    """
-    Defines an object to store metadata linked to an application
-    """
-
-    _name = 'clouder.container.metadata'
-
-    name = fields.Many2one(
-        'clouder.application.metadata', 'Application Metadata',
-        ondelete="cascade", required=True)
-    container_id = fields.Many2one(
-        'clouder.container', 'Container', ondelete="cascade", required=True)
-    value_data = fields.Text('Value')
-
-    _sql_constraints = [
-        ('name_uniq', 'unique(name, container_id)',
-         'Metadata must be unique per container!'),
-    ]
-
-    @property
-    def value(self):
-        """
-        Property that returns the value formatted by type
-        """
-        def _missing_function():
-            # If the function is missing, raise an exception
-            self.raise_error(
-                'Invalid function name "%s" for clouder.container',
-                self.name.func_name,
-            )
-
-        # Computing the function if needed
-        val_to_convert = self.value_data
-        if self.name.is_function:
-            val_to_convert = "{0}".format(getattr(
-                self.container_id, self.name.func_name, _missing_function)())
-            # If it is a function,
-            # the text version should be updated for display
-            self.with_context(skip_check=True).write({
-                'value_data': val_to_convert})
-
-        # Empty value
-        if not val_to_convert:
-            return False
-
-        # value_type cases
-        if self.name.value_type == 'int':
-            return int(val_to_convert)
-        elif self.name.value_type == 'float':
-            return float(val_to_convert)
-        # Defaults to char
-        return str(val_to_convert)
-
-    @api.multi
-    @api.constrains('name')
-    def _check_clouder_type(self):
-        """
-        Checks that the metadata is intended for containers
-        """
-        if self.name.clouder_type != 'container':
-            self.raise_error(
-                "This metadata is intended for %s only.",
-                self.name.clouder_type,
-            )
-
-    @api.multi
-    @api.constrains('name', 'value_data')
-    def _check_object(self):
-        """
-        Checks if the data can be loaded properly
-        """
-        if 'skip_check' in self.env.context and self.env.context['skip_check']:
-            return
-        # call the value property to see if the metadata can be loaded properly
-        try:
-            self.value
-        except ValueError:
-            # User display
-            self.raise_error(
-                'Invalid value for type "%s": \n\t"%s"\n',
-                self.name.value_type, self.value_data,
-            )
