+++ clouder/models/container_option.py
@@ -1,0 +1,43 @@
+# -*- coding: utf-8 -*-
+# Copyright 2015 Clouder SASU
+# License LGPL-3.0 or later (http://gnu.org/licenses/lgpl.html).
+
+from openerp import models, fields, api
+
+
+import logging
+_logger = logging.getLogger(__name__)
+
+
+class ClouderContainerOption(models.Model):
+    """
+    Define the container.option object, used to define custom values
+    specific to a container.
+    """
+
+    _name = 'clouder.container.option'
+
+    container_id = fields.Many2one(
+        'clouder.container', 'Container', ondelete="cascade", required=True)
+    name = fields.Many2one(
+        'clouder.application.type.option', 'Option', required=True)
+    value = fields.Text('Value')
+
+    _sql_constraints = [
+        ('name_uniq', 'unique(container_id,name)',
+         'Option name must be unique per container!'),
+    ]
+
+    @api.multi
+    @api.constrains('container_id')
+    def _check_required(self):
+        """
+        Check that we specify a value for the option
+        if this option is required.
+        """
+        if self.name.required and not self.value:
+            self.raise_error(
+                'You need to specify a value for the option '
+                '"%s" for the container "%s".',
+                self.name.name, self.container_id.name,
+            )
