+++ clouder/models/container_volume.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+# Copyright 2015 Clouder SASU
+# License LGPL-3.0 or later (http://gnu.org/licenses/lgpl.html).
+
+from openerp import fields
+from openerp import models
+
+
+import logging
+_logger = logging.getLogger(__name__)
+
+
+class ClouderContainerVolume(models.Model):
+    """
+    Define the container.volume object, used to define the volume which
+    will be saved in the container or will be linked to a directory
+    in the host server.
+    """
+
+    _name = 'clouder.container.volume'
+
+    container_id = fields.Many2one(
+        'clouder.container', 'Container', ondelete="cascade", required=True)
+    name = fields.Char('Path', required=True)
+    hostpath = fields.Char('Host path')
+    user = fields.Char('System User')
+    readonly = fields.Boolean('Readonly?')
+    nosave = fields.Boolean('No save?')
+
+    _sql_constraints = [
+        ('name_uniq', 'unique(container_id,name)',
+         'Volume name must be unique per container!'),
+    ]
